Reorganize servertab's code.
